requirements.txt
asyncio
jsons
psutil
python-dotenv
websockets